Adding test for error in the API response
